# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Job": {
        "description": "Full record \u2014 the source of truth persisted as `jobs/<id>.json`.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_volume_key": {
            "title": "Input Volume Key",
            "type": "string"
          },
          "mask_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mask Key"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "modality": {
            "enum": [
              "CT",
              "MRI",
              "Other"
            ],
            "title": "Modality",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "overlay_slice_keys": {
            "items": {
              "type": "string"
            },
            "title": "Overlay Slice Keys",
            "type": "array"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "thumbnail_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Key"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "input_volume_key",
          "modality",
          "model",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "Job",
        "type": "object"
      },
      "JobCreate": {
        "description": "Form fields for creating a job. Inputs are immutable once set.",
        "properties": {
          "input_volume_key": {
            "minLength": 1,
            "title": "Input Volume Key",
            "type": "string"
          },
          "modality": {
            "default": "CT",
            "enum": [
              "CT",
              "MRI",
              "Other"
            ],
            "title": "Modality",
            "type": "string"
          },
          "model": {
            "default": "demo-lesion",
            "title": "Model",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "patient_id": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "site_id": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "name",
          "input_volume_key"
        ],
        "title": "JobCreate",
        "type": "object"
      },
      "JobMetrics": {
        "description": "Quantitative outcome of one completed inference run.",
        "properties": {
          "device": {
            "title": "Device",
            "type": "string"
          },
          "foreground_voxels": {
            "title": "Foreground Voxels",
            "type": "integer"
          },
          "labels": {
            "items": {
              "$ref": "#/components/schemas/LabelVolume"
            },
            "title": "Labels",
            "type": "array"
          },
          "shape": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shape"
          },
          "spacing": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spacing"
          }
        },
        "required": [
          "device",
          "foreground_voxels"
        ],
        "title": "JobMetrics",
        "type": "object"
      },
...
      "JobSummary": {
        "description": "Compact record for the Segmentations table (one per jobs/<id>.json).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_volume_key": {
            "title": "Input Volume Key",
            "type": "string"
          },
          "modality": {
            "enum": [
              "CT",
              "MRI",
              "Other"
            ],
            "title": "Modality",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "thumbnail_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Key"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "input_volume_key",
          "modality",
          "model",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "JobSummary",
        "type": "object"
      },
      "JobUpdate": {
        "description": "Editable metadata ONLY. Inputs (volume/model/modality) are immutable by\ndesign \u2014 a job records one inference on one volume; to change the input,\ncreate a new job. All fields optional; only provided ones change.",
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "title": "JobUpdate",
        "type": "object"
      },
      "LabelVolume": {
        "description": "Physical volume of one segmented label.",
        "properties": {
          "label": {
            "title": "Label",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "volume_ml": {
            "title": "Volume Ml",
            "type": "number"
          },
          "voxels": {
            "title": "Voxels",
            "type": "integer"
          }
        },
        "required": [
          "label",
          "name",
          "voxels",
          "volume_ml"
        ],
        "title": "LabelVolume",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local API for nnU-Net 3D Medical Image Segmentation: ingest 3D CT/MRI volumes, run real nnU-Net inference on-device, and archive every artifact (raw volumes, preprocessed tensors, masks, and the model checkpoint) on Backblaze B2 over the S3-compatible API. This contract documents the local API, not a hosted public endpoint.",
    "title": "nnU-Net 3D Medical Image Segmentation API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs": {
      "get": {
        "operationId": "list_jobs_endpoint_jobs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobSummary"
                  },
                  "title": "Response List Jobs Endpoint Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Jobs Endpoint",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "operationId": "create_job_endpoint_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job Endpoint",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/jobs/{job_id}": {
      "delete": {
        "operationId": "delete_job_endpoint_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Job Endpoint",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "operationId": "get_job_endpoint_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Endpoint",
        "tags": [
          "jobs"
        ]
      },
      "patch": {
        "operationId": "update_job_endpoint_jobs__job_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Job Endpoint",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/run": {
      "post": {
        "operationId": "run_job_endpoint_jobs__job_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Job Endpoint",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/slices/{index}": {
      "get": {
        "operationId": "job_slice_endpoint_jobs__job_id__slices__index__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "schema": {
              "title": "Index",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Slice Endpoint",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/volumes/slice": {
      "get": {
        "operationId": "volume_slice_endpoint_volumes_slice_get",
        "parameters": [
          {
            "description": "Object key under volumes/ or masks/",
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "description": "Object key under volumes/ or masks/",
              "title": "Key",
              "type": "string"
            }
          },
          {
            "description": "Axial slice index; default mid-slice",
            "in": "query",
            "name": "index",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Axial slice index; default mid-slice",
              "title": "Index"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/png": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Volume Slice Endpoint",
        "tags": [
          "volumes"
        ]
      }
    }
  }
}